Initial commit.
Parsing the King James Version of the Bible from Project
Gutenberg to JSON and Minified JSON.

To do so yourself run book-parser.js and it should do everything for
you.

You will need to update the files here if you plan to use a more recent
version of the text in case anything changes.