For real this time.
This should actually be the proper data this time.

There are a bunch of weird formatting and little things within the
document and with it being 10s of thousands of lines I don't really have
the ability to see all the errors.

This however does look significantly more complete than the last deal I
had going on here.

Closes #1