Word Occurrence Total
/api/totaloccur/{word}

Returns the total number of times the supplied word is found within the
old and new testament(s). If both values are present, a third value
"both" is added that displays the combined total within both books.

Adjusted the regexp within /api/firstoccur/{word} to be inline with
totaloccur with the ignore case flag.

This now allows for wierd SpElLiNgS of words within the request.

Closes #3